REFACTOR-002: extract song/routing/device invariants into ./parse/song.ts
Pure structural move, no behavior change. Moves checkSongIntegrity
(exported), checkInstrument, checkInstrumentParameters, checkDeviceChain
(module-private), and the MAX_TRACK_INSERTS/MAX_RETURN_BUSES capacity
constants out of parse.ts into their own module. parse.ts re-exports the
same symbols unchanged.

Fifth of the REFACTOR-002 stack; builds on the pack-invariants
extraction. This module is the largest remaining cohesive unit (song
integrity, device chains, and instrument parameter checks all call each
other within one invariant), so it lands as one PR rather than
splitting further.

Refs #143
